ADD union operation to SeqIntervals
ADD unit testing for above.
FIX minor bug fixes in NCBISequence and SeqIntervals

FIXME floating point error in PositionMask causes the proportions
to occasionally not add up to 1. This causes the mask to occasionally
have the wrong values, e.g. a mask of 100% random positions may have
one value that is another type of mask. Not a critical error for
training since all mask types are usually used for BERT-like model.