ADD/MODIFY/FIX `seqbert.py` cls token, masking
ADD classification token for next sequence prediction pretext task.
This uses transforms to split and permute sequences,
and adds separate `cls_regularization` loss based on
prediction of whether each sequence in the batch was permuted.

MODIFY sequence masking logic for BERT model moved from `mask.py` to
`data_transform()` and `mask_transform()` in `seqbert.py`.
This concentrates the model-specific logic in one location.

FIX remove attention masking in `torch.nn.TransformerEncoder`,
this was incorrectly removing information for masked position prediction.
Instead, only need to replace tokens in the source sequence with a
mask token.

REMOVE flanking sequences in `mask.py`, this is no longer necessary
as it is easier to control flanking sequences in the data preprocessing
steps of the model itself.